backend/requirements.txt
Django>=4.2
psycopg2-binary>=2.9
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-cors-headers>=4.0
channels>=4.0
channels-redis>=4.1
celery>=5.3
redis>=5.0
pytest
drf-spectacular
pytest-django
